+++ app/Http/Controllers/RiderController.php
@@ -2,15 +2,24 @@
 
 namespace App\Http\Controllers;
 
+use Auth;
+use View;
 use App\Models\Rider;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Session;
-use DateTime;
-use View;
 
 class RiderController extends PublicController
 {
+
+	/**
+	 * Create a new controller instance.
+	 * @return void
+	 */
+	public function __construct()
+	{
+		$this->middleware('auth');
+	}
 
 	/**
 	 * Show the list of available Riders
@@ -19,50 +28,58 @@ class RiderController extends PublicController
 	 */
 	public function index( Request $request )
 	{
-		$rider = new \App\Models\Rider;
-
-		$searched = false;
-		$keyword = $request->input('keyword');
-		$gender = $request->input('gender');
-		$clubsId = $request->input('clubsId');
-		$appends = ['keyword' => $keyword, 'gender' => $gender, 'clubsId' => $clubsId];
-
-		if ( $request->isMethod('post') )
+		$loggedInUser = Auth::user();
+		if ( is_object($loggedInUser) && $loggedInUser->role == \App\Models\User::ROLE_ADMIN )
 		{
-			if ( !empty($keyword ) || !empty($gender) || !empty($clubsId) )
+			$rider = new \App\Models\Rider;
+
+			$searched = false;
+			$keyword = $request->input('keyword');
+			$gender = $request->input('gender');
+			$clubsId = $request->input('clubsId');
+			$appends = ['keyword' => $keyword, 'gender' => $gender, 'clubsId' => $clubsId];
+
+			if ( $request->isMethod('post') )
 			{
-				$searched = true;
-				$riders = $rider->search($keyword, $gender, $clubsId);
+				if ( !empty($keyword ) || !empty($gender) || !empty($clubsId) )
+				{
+					$searched = true;
+					$riders = $rider->search($keyword, $gender, $clubsId);
+				}
 			}
-		}
 
-		// If we haven't searched get all items as a paginated set
-		if ( !$searched )
-		{
-			$riders = $rider->paginate();
-		}
+			// If we haven't searched get all items as a paginated set
+			if ( !$searched )
+			{
+				$riders = $rider->paginate();
+			}
 
-		if ( count($riders) > 0 )
-		{
-			$riders->appends($appends);
-			$rv = view('riders.list')->with([
-				'riders' => $riders,
-				'rider' => $rider,
-				'keyword' => $keyword,
-				'gender' => $gender,
-				'clubsId' => $clubsId,
-				'searched' => $searched,
-			]);
+			if ( count($riders) > 0 )
+			{
+				$riders->appends($appends);
+				$rv = view('riders.list')->with([
+					'riders' => $riders,
+					'rider' => $rider,
+					'keyword' => $keyword,
+					'gender' => $gender,
+					'clubsId' => $clubsId,
+					'searched' => $searched,
+				]);
+			}
+			else
+			{
+				$rv = view('riders.nonesuch')->with([
+					'rider' => $rider,
+					'keyword' => $keyword,
+					'gender' => $gender,
+					'clubsId' => $clubsId,
+					'searched' => $searched,
+				]);
+			}
 		}
 		else
 		{
-			$rv = view('riders.nonesuch')->with([
-				'rider' => $rider,
-				'keyword' => $keyword,
-				'gender' => $gender,
-				'clubsId' => $clubsId,
-				'searched' => $searched,
-			]);
+			$rv = redirect('/')->with('error', 'You do not have permissions to access this feature.');
 		}
 		return $rv;
 	}
@@ -74,30 +91,32 @@ class RiderController extends PublicController
 	 */
 	public function create( Request $request )
 	{
-		$rv = NULL;
-
-		$rider = new \App\Models\Rider;
-
-		if ( $request->isMethod('post') )
+		$loggedInUser = Auth::user();
+		if ( is_object($loggedInUser) && $loggedInUser->role == \App\Models\User::ROLE_ADMIN )
 		{
-			$data = $request->get('Rider');
-			Rider::fillFromFilteredData($rider, $data);
+			$rv = NULL;
 
-			$rv = $rider->storeDetails($request);
-		}
+			$rider = new \App\Models\Rider;
 
-		if ( $rv === NULL )
-		{
-			$data = $request->session()->get('_old_input');
-			if ( isset($data['Rider']) && is_array($data['Rider']) )
+			if ( $request->isMethod('post') )
 			{
-				Rider::fillFromFilteredData($rider, $data['Rider']);
+				$data = $request->get('Rider');
+				Utilities::fillFromFilteredData($rider, $data);
+
+				$rv = $rider->storeDetails($request);
 			}
 
-			$rv = view('riders.details')->with([
-				'rider' => $rider,
-				'route' => route('/riders/create')
-			]);
+			if ( $rv === NULL )
+			{
+				$rv = view('riders.details')->with([
+					'rider' => $rider,
+					'route' => route('/riders/create')
+				]);
+			}
+		}
+		else
+		{
+			$rv = redirect('/')->with('error', 'You do not have permissions to access this feature.');
 		}
 		return $rv;
 	}
@@ -110,27 +129,29 @@ class RiderController extends PublicController
 	 */
 	public function details( $id, Request $request )
 	{
-		$rider = \App\Models\Rider::findOrFail($id);
-
-		if ( $request->isMethod('post') )
+		$loggedInUser = Auth::user();
+		if ( is_object($loggedInUser) && $loggedInUser->role == \App\Models\User::ROLE_ADMIN )
 		{
-			$data = $request->get('Rider');
-			Rider::fillFromFilteredData($rider, $data);
+			$rider = \App\Models\Rider::findOrFail($id);
 
-			$rv = $rider->storeDetails($request);
+			if ( $request->isMethod('post') )
+			{
+				$data = $request->get('Rider');
+				Utilities::fillFromFilteredData($rider, $data);
+
+				$rv = $rider->storeDetails($request);
+			}
+			else
+			{
+				$rv = view('riders.details')->with([
+					'rider' => $rider,
+					'route' => route('/riders/details', $rider->id)
+				]);
+			}
 		}
 		else
 		{
-			$data = $request->session()->get('_old_input');
-			if ( isset($data['Rider']) && is_array($data['Rider']) )
-			{
-				Rider::fillFromFilteredData($rider, $data['Rider']);
-			}
-
-			$rv = view('riders.details')->with([
-				'rider' => $rider,
-				'route' => route('/riders/details', $rider->id)
-			]);
+			$rv = redirect('/')->with('error', 'You do not have permissions to access this feature.');
 		}
 
 		return $rv;
@@ -144,16 +165,24 @@ class RiderController extends PublicController
 	 */
 	public function delete( $id, Request $request )
 	{
-		$rider = \App\Models\Rider::findOrFail($id);
-
-		$allowed = false;
-		if ( $allowed && $rider->delete() )
+		$loggedInUser = Auth::user();
+		if ( is_object($loggedInUser) && $loggedInUser->role == \App\Models\User::ROLE_ADMIN )
 		{
-			$rv = redirect('/riders/index')->with('success', 'The rider was successfully deleted');
+			$rider = \App\Models\Rider::findOrFail($id);
+
+			$allowed = false;
+			if ( $allowed && $rider->delete() )
+			{
+				$rv = redirect('/riders/index')->with('success', 'The rider was successfully deleted');
+			}
+			else
+			{
+				$rv = redirect('/riders/index')->with('error', 'An error occurred deleting the rider.');
+			}
 		}
 		else
 		{
-			$rv = redirect('/riders/index')->with('error', 'An error occurred deleting the rider.');
+			$rv = redirect('/')->with('error', 'You do not have permissions to access this feature.');
 		}
 
 		return $rv;
